♻️(api) reorganize experience index models
* Splitted Pydantic and SQLModel models in two files.
* Created a dedicated file for enumerations

These changes are inspired by Tiangolo. In his web-app boilerplate,
using FastAPI and SQLModel, Pydantic models are seperated from
database Schema.
